Overhaul directory structure
Added ext/ and lib/ directories. There's now a lib/fb.rb which is the
main entry point for the gem, and it loads `fb/fb_ext` for the C extension.

Set up building with `rake-compiler`.

Moved README to USAGE.txt and set up a README.md suitable for display online